+++ api/posts-microservice/src/main/java/com/socialsync/postsmicroservice/service/PostsService.java
@@ -1,21 +1,190 @@
 package com.socialsync.postsmicroservice.service;
 
+import com.google.gson.Gson;
+import com.socialsync.postsmicroservice.components.RabbitMqConnectionFactoryComponent;
+import com.socialsync.postsmicroservice.pojo.PostQueueMessage;
+import com.socialsync.postsmicroservice.pojo.enums.QueueMessageType;
 import com.socialsync.postsmicroservice.repository.PostRepository;
 import com.socialsync.postsmicroservice.interfaces.PostsServiceMethods;
 import com.socialsync.postsmicroservice.pojo.Post;
 import com.socialsync.postsmicroservice.util.exceptions.PostNotFound;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.amqp.core.AmqpTemplate;
+import org.springframework.context.annotation.Bean;
+import org.springframework.scheduling.annotation.EnableScheduling;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
 import java.time.Instant;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Optional;
+import java.util.Random;
 
 @AllArgsConstructor
 @Service
+@Slf4j
+@EnableScheduling
 public class PostsService implements PostsServiceMethods {
 
     private PostRepository repository;
+
+    private RabbitMqConnectionFactoryComponent conectionFactory;
+
+    private AmqpTemplate amqpTemplate;
+
+    private Gson gson;
+
+    @Bean
+    void initTemplate() {
+        this.amqpTemplate = conectionFactory.rabbitTemplate();
+    }
+
+    private void sendMessage(PostQueueMessage post) {
+        String json = gson.toJson(post);
+        this.amqpTemplate.convertAndSend(conectionFactory.getExchange(), conectionFactory.getRoutingKey(), json);
+    }
+
+    void deleteEverything() {
+        repository.findAll().forEach(x -> {
+            try {
+                deletePost(x.getId());
+            } catch (PostNotFound e) {
+                throw new RuntimeException(e);
+            }
+        });
+    }
+
+    static List<String> titluPostare = List.of(
+            "Orașul Ascuns",
+            "Luminile Noptii",
+            "Portale în Alt Univers",
+            "Misterul Trecutului",
+            "Călătorie Fără Destinație",
+            "Pe Următoarea Stea",
+            "În Căutarea Sensului",
+            "Paradisul Pierdut",
+            "În Umbra Realității",
+            "Poveștile Stelare",
+            "Labirintul Minții",
+            "Destinul Ascuns",
+            "Până la Marginea Lumii",
+            "Culoarea Vântului",
+            "Sufletul Cosmosului",
+            "Prizonier în Timp",
+            "Scriind Pe Filele Destinului",
+            "Rădăcinile Invisibile",
+            "Zborul Spre Necunoscut",
+            "Frumusețea Imaterială",
+            "Descoperind Culorile Sufletului",
+            "Călătorie Prin Nebuloase",
+            "Dincolo de Orizont",
+            "Portretul Inimii",
+            "În Căutarea Adevărului",
+            "În Vâltoarea Eternității",
+            "Când Steltele Plâng",
+            "Labirintul Timpului",
+            "Căutătorul de Răspunsuri",
+            "Povestea Fără Sfârșit",
+            "În Jocul Cuantelor",
+            "Călătorind Pe Râuri de Lumină",
+            "Intrând în Tărâmul Viselor",
+            "Oglinda Dincolo de Realitate",
+            "Legături Neexplorate",
+            "Povestea Uitată",
+            "Magia Nopții Albastre",
+            "Răsăritul Lumii Noi",
+            "Universul Într-o Picătură de Rouă",
+            "Explorând Misterul",
+            "În Căutarea Cunoașterii",
+            "Sub Semnul Stelelor",
+            "Aventuri În Dimensiuni Paralele",
+            "Călătorind Prin Portale",
+            "Scriind Pe Pergamentul Destinului",
+            "Zbor deasupra Norilor",
+            "Trăind În Vis",
+            "Misterul Pădurei Întunecate",
+            "Căutând Lumina Dincolo de Umbră",
+            "Rătăcind în Labyrinthul Timpului",
+            "Călătorind Cu Vântul",
+            "În Căutarea Armoniei",
+            "Pe Drumul Spre Infinit",
+            "În Spatele Oglindelor",
+            "Frumusețea Efemeră",
+            "Lumea Dincolo de Vise",
+            "Căutând Comori Îngropate",
+            "Sufletul Peregrin",
+            "Destinul în Carte Scrie",
+            "În Căutarea Sufletului Pierdut",
+            "Pe Drumul Norilor",
+            "Sămânța Magiei",
+            "Orașul Adormit",
+            "Sufletul Călător",
+            "Răsăritul Fără Apus",
+            "Calea Fără Sfârșit",
+            "Fereastra Către Infinit",
+            "În Pădurea cu Zâne",
+            "Scriind În Caietele Sufletului",
+            "Descoperind Pășunile Eterne",
+            "În Căutarea Inimii Pierdute",
+            "Aroma Stelelor",
+            "Călătorind Cu Luna",
+            "Urmărind Cometele",
+            "În Valurile Timpului",
+            "Portretul Cuvintelor",
+            "Ochiul Vârtejurilor",
+            "Sufletul Fugind de Realitate",
+            "Printre Pagini de Istorii",
+            "Călătorind Prin Galaxii",
+            "Răsăritul Dincolo de Munte",
+            "În Căutarea Pământului Promis",
+            "Labirintul Viselor",
+            "Sufletul Pustiit",
+            "Pe Aripile Gândurilor",
+            "Rătăcind Prin Universul Paralel",
+            "Lumea În Culori de Argint",
+            "Căutând Lumină în Întuneric",
+            "În Căutarea Nopților Nesfârșite",
+            "Misterul Căii Lactee",
+            "În Labirintul Sufletului",
+            "Călătorind În Vânt",
+            "În Căutarea Cetății de Nisip",
+            "Scriind Cu Stelele",
+            "Misterul Oglindelor",
+            "În Căutarea Soarelui de Mâine",
+            "Călătorind Pe Marea Eternității",
+            "Pe Drumul Neexplorat",
+            "În Spatele Porților Închise",
+            "Fereastra către Infinitul Interior",
+            "Aventura În Țara Minunilor"
+    );
+
+    @Bean
+    void populateDb() {
+        deleteEverything();
+
+        for (int i = 0;i < 100; i++) {
+            String titlu = titluPostare.get(new Random().nextInt(titluPostare.size()));
+            String continut = "Acesta este un conținut scurt pentru postarea cu titlul \"" + titlu + "\".";
+            Post post = new Post("-1", "-1", titlu, continut);
+            addPost(post);
+        }
+    }
+
+    @Bean
+    @Scheduled(fixedDelay = 5000L)
+    void newRandomPost() {
+        log.info("We have " + repository.findAll().size() + " comments");
+
+        if (repository.findAll().size() > 200)
+            deleteEverything();
+
+        String titlu = titluPostare.get(new Random().nextInt(titluPostare.size()));
+        String continut = "Acesta este un conținut scurt pentru postarea cu titlul \"" + titlu + "\".";
+        Post post = new Post("-1", "-1", titlu, continut);
+        addPost(post);
+    }
 
     @Override
     public HashMap<String, Post> fetchAllPosts() {
@@ -38,6 +207,7 @@ public class PostsService implements PostsServiceMethods {
     public void addPost(Post post)  {
         post.setTimestampCreated(Instant.now().getEpochSecond());
         repository.insert(post);
+        sendMessage(new PostQueueMessage(QueueMessageType.CREATE, post));
     }
 
     @Override
@@ -49,16 +219,25 @@ public class PostsService implements PostsServiceMethods {
                 elem.setTopicId(post.getTopicId());
                 elem.setTimestampUpdated(Instant.now().getEpochSecond());
                 repository.save(elem);
+                sendMessage(new PostQueueMessage(QueueMessageType.UPDATE, post));
                 return elem;
         }).orElseThrow(() -> {
             post.setTimestampCreated(Instant.now().getEpochSecond());
             repository.insert(post);
+            sendMessage(new PostQueueMessage(QueueMessageType.CREATE, post));
             return new PostNotFound("Post not found. Created one instead.");
         });
     }
 
     @Override
-    public void deletePost(String id) {
-        repository.deleteById(id);
+    public void deletePost(String id) throws PostNotFound {
+        Optional<Post> post = repository.findById(id);
+
+        if (post.isPresent()) {
+            repository.deleteById(id);
+            sendMessage(new PostQueueMessage(QueueMessageType.DELETE, post.get()));
+        }
+        else
+            throw new PostNotFound("Post not found.");
     }
 }
